APIs/telegram_bot/status-trello/src/conf/settings.py
import os

from dotenv import load_dotenv

load_dotenv()

TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")
TELEGRAM_WHITE_LIST = [1195845295]

TRELLO_API_URL = os.getenv("TRELLO_API_URL")
TRELLO_KEY = os.getenv("TRELLO_KEY")
TRELLO_TOKEN = os.getenv("TRELLO_TOKEN")