APIs/telegram_bot/status-trello/requirements.txt
altair==4.1.0
astor==0.8.1
astroid==2.4.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
bleach==3.1.5
blinker==1.4
boto3==1.13.13
botocore==1.16.13
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
construct==2.5.0
conversaosgi==0.1.1
cryptography==2.9.2
dbfread==2.0.7
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fdb==2.0.1
fire==0.3.1
future==0.18.2
idna==2.9
ipykernel==5.3.0
ipython==7.14.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.0
Jinja2==2.11.2
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
kagebunshin==0.0.1
lazy-object-proxy==1.4.3
loguru==0.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.3.0
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.6
notebook==6.0.3
numpy==1.18.4
packaging==20.4
pandas==1.0.3
pandas-access==0.0.1
pandocfilters==1.4.2
parso==0.7.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
psycopg2==2.8.5
ptyprocess==0.6.0
py==1.8.1
pycparser==2.20
pydeck==0.3.1
Pygments==2.6.1
pylint==2.5.2
PyMySQL==0.9.3
pyodbc==4.0.30
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.2
python-dateutil==2.8.1
python-dotenv==0.13.0
python-telegram-bot==12.8
pytz==2020.1
pyzmq==19.0.1
requests==2.23.0
s3transfer==0.3.3
Send2Trash==1.5.0
six==1.14.0
SQLAlchemy==1.3.16
streamlit==0.60.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==5.1.1
tqdm==4.46.0
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.9
validators==0.15.0
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
